requirements.txt
asgiref==3.5.2
attrs==22.1.0
cfgv==3.3.1
coverage==6.5.0
distlib==0.3.6
Django==4.1.1
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-extensions==3.2.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.1.3
filelock==3.8.0
identify==2.5.5
iniconfig==1.1.1
machine==0.0.1
nodeenv==1.7.0
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
pluggy==1.0.0
pre-commit==2.20.0
psycopg2-binary==2.9.3
py==1.11.0
pycrypto==2.6.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
redgreenunittest==0.1.1
six==1.16.0
sqlparse==0.4.2
tenacity==8.1.0
toml==0.10.2
tomli==2.0.1
virtualenv==20.16.5
